setting up blocks as map of tuples for raycasting

+++ src/block.h
@@ -12,20 +12,20 @@ enum blockType{
 
 class Block{
   public:
-    Block(float xPos, float yPos, float zPos) :x(xPos), y(yPos), z(zPos){
+    Block(int xPos, int yPos, int zPos) :x(xPos), y(yPos), z(zPos){
 
     }
 
     void insertVertices(std::vector<std::vector<float>> face, std::vector<std::vector<float>> tex){
-        for (int i = 0; i < face.size(); i++) {
-            std::vector<float> translatedVertex = face[i];
-            translatedVertex[0] += x; // x position
-            translatedVertex[1] += y; // y position
-            translatedVertex[2] += z; // z position
-            translatedVertex[3] = tex[i][0];
-            translatedVertex[4] = tex[i][1];
-            vertices.insert(vertices.end(), translatedVertex.begin(), translatedVertex.end());
-        }
+      for (int i = 0; i < face.size(); i++) {
+        std::vector<float> translatedVertex = face[i];
+        translatedVertex[0] += x; // x position
+        translatedVertex[1] += y; // y position
+        translatedVertex[2] += z; // z position
+        translatedVertex[3] = tex[i][0];
+        translatedVertex[4] = tex[i][1];
+        vertices.insert(vertices.end(), translatedVertex.begin(), translatedVertex.end());
+      }
     }
     
     void setType(blockType block){
@@ -40,9 +40,9 @@ class Block{
   
   private:
     blockType type;
-    float x;
-    float y;
-    float z;
+    int x;
+    int y;
+    int  z;
 };
 
 #endif
